+++ src/main/java/com/keepgoing/keepserver/domain/device/domain/entity/enums/DeviceStatus.java
@@ -1,4 +1,4 @@
-package com.keepgoing.keepserver.domain.device.entity.enums;
+package com.keepgoing.keepserver.domain.device.domain.entity.enums;
 
 public enum DeviceStatus {
     AVAILABLE,
