+++ src/main/java/com/keepgoing/keepserver/domain/device/domain/repository/DeviceRepository.java
@@ -1,7 +1,7 @@
-package com.keepgoing.keepserver.domain.device.repository;
+package com.keepgoing.keepserver.domain.device.domain.repository;
 
-import com.keepgoing.keepserver.domain.device.entity.Device;
-import com.keepgoing.keepserver.domain.user.entity.user.User;
+import com.keepgoing.keepserver.domain.device.domain.entity.Device;
+import com.keepgoing.keepserver.domain.user.domain.entity.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
