+++ src/main/java/com/keepgoing/keepserver/domain/device/mapper/DeviceMapper.java
@@ -1,7 +1,7 @@
 package com.keepgoing.keepserver.domain.device.mapper;
 
-import com.keepgoing.keepserver.domain.device.entity.Device;
-import com.keepgoing.keepserver.domain.device.entity.enums.DeviceStatus;
+import com.keepgoing.keepserver.domain.device.domain.entity.Device;
+import com.keepgoing.keepserver.domain.device.domain.entity.enums.DeviceStatus;
 import com.keepgoing.keepserver.domain.device.payload.response.DeviceResponseDto;
 import com.keepgoing.keepserver.domain.device.payload.request.DeviceDto;
 import org.springframework.stereotype.Component;
